chore(orca-api): move interfaces into api package
