+++ orca-api/src/main/java/com/netflix/spinnaker/orca/api/pipeline/ExecutionPreprocessor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Netflix, Inc.
+ * Copyright 2021 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,12 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.spinnaker.orca.extensionpoint.pipeline;
+package com.netflix.spinnaker.orca.api.pipeline;
 
-import java.util.Map;
 import javax.annotation.Nonnull;
+import java.util.Map;
 
-/** A preprocessor that can modify an Execution upon initial receipt of the configuration. */
+/**
+ * ExecutionPreprocessor is a hook point that can modify an Execution upon initial receipt
+ * of the configuration.
+ *
+ * <p>A common use case is to inspect and update a pipeline with configured context, such
+ * an application name.
+ */
 public interface ExecutionPreprocessor {
 
   /** Returns whether or not the preprocess can handle the inbound execution. */
